Builtins, multiple versions, and new installs
Trying to (require-package 'project '(0 5 3)) doesn't work, because 'project is builtin, and even though the version isn't sufficient, (package-install 'project) will just choose the builtin. The package-desc needs to be passed in. Along with that, I try to select the highest version.

The return value from (package-install package) is sometimes nil if the package was not previously downloaded, so I've changed the return value of (require-package) to be whether the package is actually installed. (I tested this by rm -rf the package-user-dir and then starting up emacs.)